# server/src/Tgm.Roborally.Server/Models/Direction.cs
/*
 * Robot Rally Game logic engine
 *
 * This api controlls the flow of a game and provides it's data. It is desiged to be RESTfull so the structure works simmilar as file system. The service will run and only work in a local network, `game.host` is the IP of the Computer hosting the game and will be found via a IP scan
 *
 * The version of the OpenAPI document: 0.1.0
 * Contact: nbrugger@student.tgm.ac.at
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using Tgm.Roborally.Server.Converters;

namespace Tgm.Roborally.Server.Models
{ 
        /// <summary>
        /// Defines a direction in a 2D system. the X and Y values can be defined by the client.
        /// </summary>
        /// <value>Defines a direction in a 2D system. the X and Y values can be defined by the client.</value>
        [TypeConverter(typeof(CustomEnumConverter<Direction>))]
        [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
        public enum Direction
        {
            
            /// <summary>
            /// Enum Up for up
            /// </summary>
            [EnumMember(Value = "up")]
            Up = 1,
            
            /// <summary>
            /// Enum Down for down
            /// </summary>
            [EnumMember(Value = "down")]
            Down = 2,
            
            /// <summary>
            /// Enum Left for left
            /// </summary>
            [EnumMember(Value = "left")]
            Left = 3,
            
            /// <summary>
            /// Enum Right for right
            /// </summary>
            [EnumMember(Value = "right")]
            Right = 4
        }
}