# server/src/Tgm.Roborally.Server/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "nbrugger@student.tgm.ac.at",
      "name" : "Nils Brugger"
    },
    "description" : "This api controlls the flow of a game and provides it's data.\nIt is desiged to be RESTfull so the structure works simmilar as file system.\nThe service will run and only work in a local network, `game.host` is the IP of the Computer hosting the game and will be found via a IP scan",
    "license" : {
      "name" : "Apache 2.0"
    },
    "title" : "Robot Rally Game logic engine",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "description" : "Production",
    "url" : "http://game.host/v1"
  }, {
    "description" : "Debug",
    "url" : "http://localhost:5050/v1"
  } ],
...
  "paths" : {
    "/games/" : {
      "get" : {
        "description" : "Returns a list of all hosted games",
        "operationId" : "getGames",
        "parameters" : [ {
          "description" : "true: only return joinable games",
          "explode" : true,
          "in" : "query",
          "name" : "joinable",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          },
          "style" : "form"
        }, {
          "description" : "true: only display games with no password set",
          "explode" : true,
          "in" : "query",
          "name" : "unprotected",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/GameID"
                  },
                  "minItems" : 0,
                  "type" : "array",
                  "uniqueItems" : true
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get all games",
        "tags" : [ "Game" ]
      },
      "post" : {
        "description" : "Creates a random game by your defined rules",
        "operationId" : "createGame",
        "parameters" : [ ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GameRules"
              }
            }
          },
          "description" : "*Optional*\nThis rules define how your game will behave"
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "Host-token-access" : [ ]
        } ],
        "summary" : "Create Game",
        "tags" : [ "Game" ]
      }
    },
    "/games/{game_id}/status" : {
      "get" : {
        "description" : "Returns the status of a game",
        "operationId" : "getGameState",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GameInfo"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get game status",
        "tags" : [ "Game" ]
      }
    },
    "/games/{game_id}/actions" : {
      "get" : {
        "description" : "Get all (**not robot related**) actions comitted to this game.",
        "operationId" : "getActions",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "description" : "Defines wich entries to return",
          "explode" : true,
          "in" : "query",
          "name" : "mode",
          "required" : false,
          "schema" : {
            "default" : "ALL",
            "enum" : [ "PENDING", "EXECUTED", "ALL" ],
            "type" : "string"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Action"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "201" : {
            "description" : "Created"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Host-token-access" : [ ]
        } ],
        "summary" : "Get games actions",
        "tags" : [ "Game" ]
      },
      "put" : {
        "description" : "Queues an action to be executed",
        "operationId" : "commitAction",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "action",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ActionType"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "security" : [ {
          "Host-token-access" : [ ]
        } ],
        "summary" : "Commit Action",
        "tags" : [ "Game" ]
      }
    },
...
    "/games/{game_id}/map/tiles/{x}/{y}" : {
      "get" : {
        "description" : "Inspect a tile of the map",
        "operationId" : "getTile",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "x",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "y",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Tile"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get tile",
        "tags" : [ "Map" ]
      }
    },
    "/games/{game_id}/map/info" : {
      "get" : {
        "description" : "Get meta information abouzt the map of the game",
        "operationId" : "getMapInfo",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MapInfo"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get Map info",
        "tags" : [ "Map" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Tile" : {
        "additionalProperties" : false,
        "description" : "A tile is a square at the Game field<br>\n**Note:** \n* `direction` is only aviable for belts\n* properties with a `rotator-` prefix are only aviable for rotator (parts)",
        "example" : {
          "empty" : true
        },
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/TileType"
          },
          "empty" : {
            "description" : "If there is no player at the tile this is true",
            "type" : "boolean"
          },
          "direction" : {
            "$ref" : "#/components/schemas/Direction"
          },
          "rotator-direction" : {
            "$ref" : "#/components/schemas/Rotation"
          }
        },
        "required" : [ "type" ],
        "title" : "Tile",
        "type" : "object"
      },
      "TileType" : {
        "description" : "The type defines the function of a tile and how it is displayed. The behaviour is unknown to the client",
        "enum" : [ "normal", "rotator", "wall", "prio_core", "conveyor", "trap-door", "stomper", "radioactive", "repair site", "Button", "One Way Wall", "Puddle", "pit", "Ramp" ],
        "title" : "TileType",
        "type" : "string"
      },
      "Direction" : {
        "default" : "up",
        "description" : "Defines a direction in a 2D system. the X and Y values can be defined by the client.",
        "enum" : [ "up", "down", "left", "right" ],
        "title" : "Direction",
        "type" : "string"
      },
...
      "Rotation" : {
        "default" : "left",
        "description" : "Defines wether to turn left or right",
        "enum" : [ "left", "right" ],
        "title" : "Rotation",
        "type" : "string"
      },
...
      "GameRules" : {
        "description" : "Game Rules define the rules for a Game",
        "example" : {
          "password" : "QWer7zÜ",
          "max-players" : 1,
          "name" : "name",
          "player-names-visible" : true,
          "fill-with-bots" : false,
          "robots-per-player" : 2
        },
        "properties" : {
          "player-names-visible" : {
            "default" : true,
            "description" : "If true players can see the name of the player controlling a robot",
            "type" : "boolean"
          },
          "max-players" : {
            "default" : 4,
            "description" : "The maximum ammount of players able to join the game",
            "maximum" : 10,
            "minimum" : 1,
            "type" : "integer"
          },
          "name" : {
            "description" : "The visible name of the game",
            "maxLength" : 50,
            "minLength" : 3,
            "type" : "string"
          },
          "robots-per-player" : {
            "description" : "Defines the number of robots per player",
            "format" : "int32",
            "maximum" : 3,
            "minimum" : 1,
            "type" : "integer"
          },
          "password" : {
            "description" : "The password of a game",
            "example" : "QWer7zÜ",
            "format" : "password",
            "maxLength" : 18,
            "minLength" : 4,
            "title" : "Password",
            "type" : "string"
          },
          "fill-with-bots" : {
            "default" : false,
            "description" : "If true emply player slots are going to be filled up with AI enemys",
            "type" : "boolean"
          }
        },
        "required" : [ "name" ],
        "title" : "GameRules",
        "type" : "object"
      },
...
      "GameID" : {
        "description" : "**Unique**<br>\nThis is the parameter a game is identified by",
        "example" : 537,
        "format" : "int32",
        "maximum" : 2048,
        "minimum" : 0,
        "title" : "GameID",
        "type" : "integer"
      },
...
      "GameInfo" : {
        "description" : "Describes the current state of the game. It does not includes the gamedata (like players/poitions/field) but general information.\n\nThis is read only",
        "example" : {
          "hardware-attached" : false,
          "hardware-compatible" : false,
          "player-on-turn" : 3,
          "passed-time" : 240
        },
        "properties" : {
          "passed-time" : {
            "default" : -1,
            "description" : "The time passed since the game started in secconds. If the game is not started it will be `0`",
            "example" : 240,
            "format" : "int32",
            "minimum" : -1,
            "type" : "integer"
          },
          "state" : {
            "$ref" : "#/components/schemas/GameState"
          },
          "hardware-compatible" : {
            "default" : false,
            "description" : "Not every game can be connected to hardware (for example to many bots)\n\nIf this is true it means you can use this game with hardware",
            "type" : "boolean"
          },
          "hardware-attached" : {
            "default" : false,
            "description" : "Is a hardware boead connected",
            "type" : "boolean"
          },
          "player-on-turn" : {
            "description" : "This id uniquely identifys the player (in a game).\n\n **Not** to be confused with the `uid` which is used for authentication",
            "example" : 3,
            "maximum" : 8,
            "minimum" : 0,
            "title" : "PlayerID",
            "type" : "integer"
          }
        },
        "required" : [ "hardware-attached", "hardware-compatible", "passed-time", "player-on-turn", "state" ],
        "title" : "Game Info",
        "type" : "object"
      },
      "GameState" : {
        "description" : "The phase the game is currently in\n* `Lobby`: Players are able to join, Bots able to be added. Host can decide to start the game and leave the phase\n* `Planning`: **[BETA - Not in game]** Players can choose their type of robot/bots bots autopick\n* `Playing`: The game is running and the players can do interactions\n* `Break`: The game is paused by the host. Players can still do interactions but they wont be executed as long as the break lasts\n* `Finished` : The game is over and there is a winner. This is the time to save the game for a replay\n",
        "enum" : [ "LOBBY", "PLANNING", "PLAYING", "FINISHED", "BREAK" ],
        "title" : "GameStatus",
        "type" : "string"
      },
...
      "Error-Message" : {
        "description" : "A message about that went wrong. Usefull to display users a short and usefull prompt",
        "properties" : {
          "message" : {
            "description" : "A short message describing what happened in human words",
            "type" : "string"
          },
          "error" : {
            "description" : "The error/exception",
            "type" : "string"
          }
        },
        "required" : [ "message" ],
        "title" : "Error-Message",
        "type" : "object"
      },
      "Action" : {
        "description" : "A queued action.\nActions are executed in their adding sequence which is represented by their index",
        "example" : {
          "index" : 0,
          "executed" : true,
          "requestor" : 6
        },
        "properties" : {
          "index" : {
            "description" : "The queue index of the action",
            "type" : "integer"
          },
          "type" : {
            "$ref" : "#/components/schemas/ActionType"
          },
          "executed" : {
            "description" : "true if the action was allready executed",
            "type" : "boolean"
          },
          "requestor" : {
            "description" : "The index of the player this instruction came from",
            "type" : "integer"
          }
        },
        "title" : "Action",
        "type" : "object"
      },
...
      "ActionType" : {
        "description" : "Defines what an action will do",
        "enum" : [ "PAUSE", "UNPAUSE", "START_GAME" ],
        "title" : "ActionType",
        "type" : "string"
      },
...
      "Position" : {
        "description" : "Describes a point in a 2D system, without decimals",
        "example" : {
          "x" : 1,
          "y" : 5
        },
        "properties" : {
          "x" : {
            "description" : "The X cordinate",
            "multipleOf" : 1,
            "type" : "integer"
          },
          "y" : {
            "description" : "Position on the Y axis (top-down on screen)",
            "multipleOf" : 1,
            "type" : "integer"
          }
        },
        "required" : [ "x", "y" ],
        "title" : "Position",
        "type" : "object"
      },
      "MapInfo" : {
        "description" : "Describes the map *without* the tiles",
        "example" : {
          "width" : 43,
          "prioBeacon" : {
            "x" : 1,
            "y" : 5
          },
          "height" : 302
        },
        "properties" : {
          "width" : {
            "maximum" : 500,
            "minimum" : 4,
            "type" : "integer"
          },
          "height" : {
            "maximum" : 500,
            "minimum" : 4,
            "type" : "integer"
          },
          "prioBeacon" : {
            "$ref" : "#/components/schemas/Position"
          }
        },
        "title" : "MapInfo",
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "Host-token-access" : {
        "description" : "The Host of the game (and of the API) has an unique identificaton granting him game manipulating abilities. By this means it is also possible to have a host and a client at the same machine. \n> This could be used for AI`s (bot players)",
        "in" : "query",
        "name" : "hid",
        "type" : "apiKey"
      },
      "Player-Access-Token" : {
        "description" : "Players need a token to join a lobby/game. If the host sets a password the player needs the password to get the token",
        "in" : "query",
        "name" : "pat",
        "type" : "apiKey"
      }
    }
  }
}